Develop

+++ src/main/java/com/ecommerce/repositories/impl/OrderRepositoryImpl.java
@@ -1,10 +1,10 @@
 package com.ecommerce.repositories.impl;
 
 import com.ecommerce.repositories.OrderRepository;
+import com.ecommerce.repositories.entites.CategoryEntity;
 import com.ecommerce.repositories.entites.OrderEntity;
 
-public class OrderRepositoryImpl  extends RepositoryImpl<OrderEntity, Long> implements OrderRepository {
-
+public class OrderRepositoryImpl extends RepositoryImpl<OrderEntity, Integer> implements OrderRepository {
     private static final OrderRepositoryImpl INSTANCE = new OrderRepositoryImpl();
 
     private OrderRepositoryImpl() {
@@ -14,6 +14,4 @@ public class OrderRepositoryImpl  extends RepositoryImpl<OrderEntity, Long> impl
     public static OrderRepositoryImpl getInstance() {
         return INSTANCE;
     }
-
-
 }
